Make Behat passing with the new directory structure
