+++ .env.test.dist
@@ -3,7 +3,7 @@
 # https://symfony.com/doc/current/best_practices/configuration.html#infrastructure-related-configuration
 
 ###> symfony/framework-bundle ###
-APP_ENV=dev
+APP_ENV=test
 APP_DEBUG=1
 APP_SECRET=EDITME
 ###< symfony/framework-bundle ###
@@ -19,5 +19,5 @@ DATABASE_URL=mysql://root@127.0.0.1/sylius_${APP_ENV}?serverVersion=5.5
 # For Gmail as a transport, use: "gmail://username:password@localhost"
 # For a generic SMTP server, use: "smtp://localhost:25?encryption=&auth_mode="
 # Delivery is disabled by default via "null://localhost"
-MAILER_URL=smtp://localhost
+MAILER_URL=null://localhost
 ###< symfony/swiftmailer-bundle ###
