+++ .travis.yml
@@ -53,6 +53,9 @@ cache:
         - $SYLIUS_CACHE_DIR
 
 before_install:
+    - mv .env.test.dist .env.test
+    - set -a && source .env.test && set +a
+
     - etc/travis/run-suite before_install "${SYLIUS_SUITE}"
 
 install:
