+++ etc/travis/suites/application/before_script.sh
@@ -9,8 +9,8 @@ run_command "bin/console cache:warmup --env=test_cached --no-debug -vvv" || exit
 run_command "bin/console doctrine:migrations:migrate --no-interaction --env=test_cached --no-debug -vvv" || exit $?
 
 print_header "Setting the web assets up" "Sylius"
-run_command "bin/console assets:install web --env=test_cached --no-debug -vvv" || exit $?
+run_command "bin/console assets:install --env=test_cached --no-debug -vvv" || exit $?
 run_command "yarn build" || exit $?
 
 print_header "Making filesystem readonly" "Sylius"
-run_command "chmod -R 555 app bin docs features src tests vendor"
+run_command "chmod -R 555 app bin config docs features src templates tests vendor"
