+++ phpunit.xml.dist
@@ -13,7 +13,7 @@
     <php>
         <ini name="error_reporting" value="-1" />
 
-        <server name="KERNEL_CLASS" value="TestKernel" />
+        <server name="KERNEL_CLASS" value="Kernel" />
         <server name="APP_ENV" value="test" />
         <server name="APP_DEBUG" value="1" />
         <server name="APP_SECRET" value="SyLiUsIsAwEsOmE" />
