+++ src/Sylius/Behat/Resources/config/profiles/default.yml
@@ -41,10 +41,11 @@ default:
             show_auto: false
 
         FriendsOfBehat\SymfonyExtension:
+            env_file: .env.test
             kernel:
-                class: TestAppKernel
-                path: app/TestAppKernel.php
-                bootstrap: vendor/autoload.php
+                class: Kernel
+                path: src/Kernel.php
+                bootstrap: ~
 
         FriendsOfBehat\ContextServiceExtension:
             imports:
